Move register + auto-login orchestration to Cloud
Relocate register_and_auto_login, the background retry loop, and cancel_auto_login from CognitoAuth to Cloud, alongside login/logout. The loop now calls self.login() and self.auth.async_register() rather than reaching back up into Cloud from the auth client, matching the existing _subscription_reconnection_task pattern. Public API is now cloud.register_and_auto_login() / cloud.cancel_auto_login().

The backoff constants move to the top-level const.py, and the tests move to test_init.py. CloudConnectionError is re-exported at the top level for consistency with the other auth exceptions. No behaviour change.

The auth package (cognito.py, auth/const.py) and tests/test_auth.py are reverted to their original state.
