+++ hass_nabucasa/const.py
@@ -14,6 +14,10 @@ TWELVE_HOURS_IN_SECONDS = 12 * 60 * 60
 ONE_HOUR_IN_SECONDS = 60 * 60
 FIVE_MINUTES_IN_SECONDS = 5 * 60
 
+# Auto-login retry backoff (register + auto-login after confirmation).
+AUTO_LOGIN_INITIAL_BACKOFF = 5  # seconds; delay before the first retry
+AUTO_LOGIN_MAX_TOTAL_BACKOFF = 86400  # seconds; give up after ~1 day of waiting
+
 MODE_PROD = "production"
 MODE_DEV = "development"
