+++ tests/test_auth.py
@@ -1,7 +1,6 @@
 """Tests for the tools to communicate with the cloud."""
 
 import asyncio
-import logging
 import time
 from unittest.mock import AsyncMock, MagicMock, patch
 
@@ -10,10 +9,6 @@ from pycognito.exceptions import MFAChallengeException
 import pytest
 
 from hass_nabucasa import CloudError, auth as auth_api
-from hass_nabucasa.auth.const import (
-    AUTO_LOGIN_INITIAL_BACKOFF,
-    AUTO_LOGIN_MAX_TOTAL_BACKOFF,
-)
 from tests.common import FROZEN_NOW_AS_TIMESTAMP
 
 
@@ -358,197 +353,3 @@ async def test_sleep_time_calculation(
         await auth._async_handle_token_refresh()
 
         assert f"Sleeping for {expected_sleep} before refreshing token" in caplog.text
-
-
-async def test_register_and_auto_login_logs_in_after_confirmation(
-    mock_cognito,
-    mock_cloud,
-):
-    """Test auto login retries until the account is confirmed, then logs in."""
-    auth = auth_api.CognitoAuth(mock_cloud)
-    mock_cloud.login = AsyncMock(side_effect=[auth_api.UserNotConfirmed(), None])
-
-    with patch("hass_nabucasa.auth.cognito.asyncio.sleep", AsyncMock()) as mock_sleep:
-        await auth.async_register_and_auto_login(
-            "email@home-assistant.io",
-            "password",
-            client_metadata={"test": "metadata"},
-        )
-        task = auth._auto_login_task
-        assert task is not None
-        await task
-
-    assert len(mock_cognito.register.mock_calls) == 1
-    assert mock_cloud.login.call_count == 2
-    mock_cloud.login.assert_called_with("email@home-assistant.io", "password")
-    assert mock_sleep.call_count == 1
-    assert auth._auto_login_task is None
-
-
-async def test_register_and_auto_login_retries_transient_errors(
-    mock_cognito,
-    mock_cloud,
-):
-    """Test auto login retries transient connection and timeout errors."""
-    auth = auth_api.CognitoAuth(mock_cloud)
-    mock_cloud.login = AsyncMock(
-        side_effect=[
-            auth_api.CloudConnectionError(),
-            auth_api.AuthTimeoutError("timeout"),
-            None,
-        ],
-    )
-
-    with patch("hass_nabucasa.auth.cognito.asyncio.sleep", AsyncMock()):
-        await auth.async_register_and_auto_login("email@home-assistant.io", "password")
-        task = auth._auto_login_task
-        assert task is not None
-        await task
-
-    assert mock_cloud.login.call_count == 3
-    assert auth._auto_login_task is None
-
-
-async def test_register_and_auto_login_stops_on_fatal_error(
-    mock_cognito,
-    mock_cloud,
-):
-    """Test auto login stops immediately on a non-retryable error."""
-    auth = auth_api.CognitoAuth(mock_cloud)
-    mock_cloud.login = AsyncMock(side_effect=auth_api.Unauthenticated("nope"))
-
-    with patch("hass_nabucasa.auth.cognito.asyncio.sleep", AsyncMock()) as mock_sleep:
-        await auth.async_register_and_auto_login("email@home-assistant.io", "password")
-        task = auth._auto_login_task
-        assert task is not None
-        await task
-
-    assert mock_cloud.login.call_count == 1
-    assert mock_sleep.call_count == 0
-    assert auth._auto_login_task is None
-
-
-async def test_register_and_auto_login_gives_up_after_one_day(
-    mock_cognito,
-    mock_cloud,
-):
-    """Test auto login backs off exponentially and gives up after ~1 day."""
-    auth = auth_api.CognitoAuth(mock_cloud)
-    mock_cloud.login = AsyncMock(side_effect=auth_api.UserNotConfirmed())
-
-    with patch("hass_nabucasa.auth.cognito.asyncio.sleep", AsyncMock()) as mock_sleep:
-        await auth.async_register_and_auto_login("email@home-assistant.io", "password")
-        task = auth._auto_login_task
-        assert task is not None
-        await task
-
-    sleeps = [call.args[0] for call in mock_sleep.mock_calls]
-    assert sleeps
-    # Delays double each time, capped at the total budget.
-    for index, value in enumerate(sleeps):
-        assert value == min(
-            AUTO_LOGIN_INITIAL_BACKOFF * 2**index,
-            AUTO_LOGIN_MAX_TOTAL_BACKOFF,
-        )
-    # The accumulated wait never exceeds the one-day budget...
-    assert sum(sleeps) <= AUTO_LOGIN_MAX_TOTAL_BACKOFF
-    # ...and it gave up because the next delay would have exceeded it.
-    next_backoff = min(
-        AUTO_LOGIN_INITIAL_BACKOFF * 2 ** len(sleeps),
-        AUTO_LOGIN_MAX_TOTAL_BACKOFF,
-    )
-    assert sum(sleeps) + next_backoff > AUTO_LOGIN_MAX_TOTAL_BACKOFF
-    # One final attempt happens after the last sleep, before giving up.
-    assert mock_cloud.login.call_count == len(sleeps) + 1
-    assert auth._auto_login_task is None
-
-
-async def test_register_and_auto_login_handles_concurrent_login_race(
-    mock_cognito,
-    mock_cloud,
-):
-    """Test a login winning the race with async_login's assert is a no-op."""
-    auth = auth_api.CognitoAuth(mock_cloud)
-
-    async def racing_login(*args, **kwargs):
-        """Simulate another task logging in just before async_login's assert."""
-        mock_cloud.is_logged_in = True
-        raise AssertionError("Cannot login if already logged in.")
-
-    mock_cloud.login = AsyncMock(side_effect=racing_login)
-
-    with patch("hass_nabucasa.auth.cognito.asyncio.sleep", AsyncMock()) as mock_sleep:
-        await auth.async_register_and_auto_login("email@home-assistant.io", "password")
-        task = auth._auto_login_task
-        assert task is not None
-        await task
-
-    assert mock_cloud.login.call_count == 1
-    assert mock_sleep.call_count == 0
-    assert auth._auto_login_task is None
-
-
-async def test_register_and_auto_login_register_failure_short_circuits(
-    mock_cognito,
-    mock_cloud,
-):
-    """Test a failed registration propagates and never starts auto login."""
-    auth = auth_api.CognitoAuth(mock_cloud)
-    mock_cognito.register.side_effect = aws_error("UsernameExistsException")
-    mock_cloud.login = AsyncMock()
-
-    with pytest.raises(auth_api.UserExists):
-        await auth.async_register_and_auto_login("email@home-assistant.io", "password")
-
-    assert auth._auto_login_task is None
-    assert mock_cloud.login.call_count == 0
-
-
-async def test_cancel_auto_login(mock_cognito, mock_cloud):
-    """Test cancelling a pending auto login stops the retry loop."""
-    auth = auth_api.CognitoAuth(mock_cloud)
-    started = asyncio.Event()
-    parked = asyncio.Event()
-
-    async def blocking_login(*args, **kwargs):
-        """Park in the first login attempt until cancelled."""
-        started.set()
-        await parked.wait()
-
-    mock_cloud.login = AsyncMock(side_effect=blocking_login)
-
-    await auth.async_register_and_auto_login("email@home-assistant.io", "password")
-    task = auth._auto_login_task
-    assert task is not None
-
-    await started.wait()
-    auth.cancel_auto_login()
-
-    assert auth._auto_login_task is None
-    with pytest.raises(asyncio.CancelledError):
-        await task
-    assert mock_cloud.login.call_count == 1
-
-
-async def test_register_and_auto_login_does_not_retain_credentials(
-    mock_cognito,
-    mock_cloud,
-    caplog,
-):
-    """Test credentials are never stored on the instance or logged."""
-    password = "sup3r-s3cr3t-p4ssw0rd"
-    auth = auth_api.CognitoAuth(mock_cloud)
-    mock_cloud.login = AsyncMock(side_effect=[auth_api.UserNotConfirmed(), None])
-
-    with (
-        caplog.at_level(logging.DEBUG, logger="hass_nabucasa.auth.cognito"),
-        patch("hass_nabucasa.auth.cognito.asyncio.sleep", AsyncMock()),
-    ):
-        await auth.async_register_and_auto_login("email@home-assistant.io", password)
-        task = auth._auto_login_task
-        assert task is not None
-        await task
-
-    assert auth._auto_login_task is None
-    assert all(value != password for value in vars(auth).values())
-    assert password not in caplog.text
